Improvements and added Textures
I have implemented all the textures provided by Sami for the
bricks/blocks. I have added in a random number that is calculated when
the world is generated from the source image. This is in a range of
0-100. Depending on the number the block will be set as a specific type
of block. Some blocks have a higher chance than others of being spawn.
For example the grass block has a 70% chance of being spawned while a
crystal resource block only has 5% chance.