100% Tested and Complete
The new resources system is now fully tested and works for both clients
and the server. Other bugs such as all players retrieving resources when
1 player does has been resolved by changing where the key press is
checked. Bugs with not being able to attach to non-default blocks has
also been resolved by adjusting the find nearest block function. This is
the fully implemented version of taking resources.